[TEP-0137] Rename customrun package
The pkg/reconciler/customrun package contains a controller that
today sends cloudevents for CustomRuns. In the context of TEP-0137,
more controllers will be added to send events for all resources,
so the current name does not make sense anymore.

Renaming the package to "notifications/customrun", so that the
notifications package may hold the various controllers for the
different kind of resources. The "events" and "cloudevents" names
are already used elsewhere and we don't want a mass API rename
for this. We might consolidate the code further as a follow-up.
The new name is also more future-proof, as the controllers may
in future handle other kinds of non-cloudevents type of notifications.

When renaming I realised that the customrun controller was being
started as part of the main controller as well as an independent
one. This PR fixes that bug as well.
