+++ cmd/controller/main.go
@@ -26,7 +26,6 @@ import (
 
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
-	"github.com/tektoncd/pipeline/pkg/reconciler/customrun"
 	"github.com/tektoncd/pipeline/pkg/reconciler/pipelinerun"
 	"github.com/tektoncd/pipeline/pkg/reconciler/resolutionrequest"
 	"github.com/tektoncd/pipeline/pkg/reconciler/taskrun"
@@ -131,7 +130,6 @@ func main() {
 		taskrun.NewController(opts, clock.RealClock{}, tpTaskrun),
 		pipelinerun.NewController(opts, clock.RealClock{}, tpPipelineRun),
 		resolutionrequest.NewController(clock.RealClock{}),
-		customrun.NewController(),
 	)
 
 	// Cleanly shutdown and flush telemetry when the application exits.
